Install necessary dependencies on Linux
Several important dependencies get installed in `homebrew.sh` with brew
on OSX, and we forgot to install them on Linux. Change the script and its
documentation so that Linux users are instructed to run `homebrew.sh` even
though it won't install brew for them. Instead, it will just install the
dependencies that would have been installed with brew.