Make path bold in default terminal prompt
This is a first step towards coming up with a color scheme that we all
like for the default terminal prompt. This makes it a bit easier to pick
out the prompt from a terminal full of text without having to worry about
different terminal background colors/colorblindness.

Partially address #6.