Fixed Windows text editor apps hogging the WSL bash standard out.
When Windows text editors such as Atom are run from WSL using `atom`
for example, they lock up the bash standard out as they do not open
in the background, and when you make certain changes/actions in the
text editor, random stuff is loged to the standard out. With this
change, text editor commands stored in `windows_apps` are aliased
to avoid this behavior. By default, commands for Sublime Text, Atom,
and Visual Studio Code are already included.