Adjust Atom settings on WSL to set default line endings to LF
On Windows, Atom installs with CRLF line endings set as the default
in Atom. This makes files sometimes incompatible for running with WSL
Bash (or any bash) as they expect LF file endings. This outght to patch
the issue, and if there is ever still and issues, you can use run
`d2u <filename>` on any file which produces some sort of Illegal
Character error when tryint to run it from WSL bash, and it will be
converted to LF line endings.