Added rubocop install, and fixed rubocop linter setting, + default line endings in WSL.
Rubocop installed by `gem install rubocop` is in a shimmed rbenv dir, so
this dir was added to the atom config using init.coffee. On WSL, rubocop
is a WSL binary wheeras Atom is a Windows binary, so this change makes a
.bat file to wrap the WSL binary so that it can be accessed from Atom.
On WSL, these changes also set the default line-endings for the editor
to LF instead of the Windows CRLF default.