Clarify instructions re: Homebrew
As Arjun pointed out in #63, the instructions weren't super clear here.
Updated them to make it clear that the script must be run regardless of OS.
Luckily, the script already checks which OS you're on, so there's no harm in
running both versions of it on OSX and Ubuntu.

Close #63.